lib: avoid return value overflow from feed_w3m
If w3m were to output more than 2 GB of plain text, the feed_w3m()
return value would overflow on 32-bit architectures and the result be
interpreted as w3m failing. Not that it matters — a 32-bit process
would have a hard time storing the 2 GB output from the pipe in
memory anyway to begin with…

+++ include/gromox/fileio.h
@@ -80,7 +80,7 @@ class GX_EXPORT wrapfd {
 
 extern GX_EXPORT std::string iconvtext(const char *, size_t, const char *from, const char *to);
 extern GX_EXPORT pid_t popenfd(const char *const *, int *, int *, int *, const char *const *);
-extern GX_EXPORT ssize_t feed_w3m(const void *in, size_t insize, std::string &out);
+extern GX_EXPORT int feed_w3m(const void *in, size_t insize, std::string &out);
 extern GX_EXPORT std::vector<std::string> gx_split(const std::string_view &, char sep);
 extern GX_EXPORT DIR_mp opendir_sd(const char *, const char *);
 extern GX_EXPORT std::unique_ptr<FILE, file_deleter> fopen_sd(const char *, const char *);
